A bunch more warning fixes... mostly conservative.

+++ libsrc/restructure.h
@@ -2,10 +2,10 @@
  * \file restructure.h
  * \brief Just declares the prototype of restructure_array().
  */
-extern void restructure_array(int ndims,
+extern void restructure_array(size_t ndims,
                               unsigned char *array, 
                               const size_t *lengths_perm,
-                              int el_size,
+                              size_t el_size,
                               const int *map,
                               const int *dir);
